refactor: move components from app directory
