+++ src/app/components/pages/QA/Question/style.module.css
